Skip setup on --control for commands with no --branch arg
The assumption is that for commands that don't have a branch arg (django-manage, ssh, etc.)
the branch it's run on doesn't matter that much: that's why it doesn't use the --branch mechanism
to force the user to think about what branch they're on. For such commands, then,
pulling the latest code on the control machine and re-running the install steps
is an unneccessary waste of time.

Commands that do have the --branch option ('ansible-playbook', 'deploy', etc.)
have essentially declared that they are sensitive to the latest changes.
These commands will continue to trigger --control to do a full reset on the control environment
as it has been doing.

+++ src/commcare_cloud/commcare_cloud.py
@@ -104,14 +104,23 @@ def run_on_control_instead(args, sys_argv):
     argv = sys_argv[1:]
     argv.remove('--control')
     executable = 'commcare-cloud'
+    force_latest_code = hasattr(args, 'branch')
     branch = getattr(args, 'branch', 'master')
     default_virtualenv = "ansible" if sys.version_info[0] == 2 else "cchq"
     venv = os.environ.get("CCHQ_VIRTUALENV", default_virtualenv)
+    if force_latest_code:
+        bash_commands_template = (
+            '{env}cd ~/commcare-cloud && git fetch --prune && git checkout {branch} '
+            '&& git reset --hard origin/{branch} && source ~/init-ansible && {cchq} {cchq_args}'
+        )
+    else:
+        bash_commands_template = (
+            '{env}cd ~/commcare-cloud && source ~/.virtualenvs/${{CCHQ_VIRTUALENV:-cchq}}/bin/activate '
+            '&& {cchq} {cchq_args}'
+        )
     cmd_parts = [
         executable, args.env_name, 'ssh', 'control[0]', '-t',
-        '{env}cd ~/commcare-cloud && git fetch --prune && git checkout {branch} '
-        '&& git reset --hard origin/{branch} && source ~/init-ansible && {cchq} {cchq_args}'
-        .format(
+        bash_commands_template.format(
             env=('export CCHQ_VIRTUALENV=%s; ' % venv if venv else ''),
             branch=branch,
             cchq=executable,
